Updates boids_sprite_sim.py
* Moves flock based rules into flock_rules.py
* Creates single function to apply all flock rules at once
* Updates references in the sim
